requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.4
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
Pillow==10.0.0
psycopg2==2.9.7
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
